Add TimeSeriesClustering
Add TimeSeriesClustering.jl package to Energy System Assessment - Optimization. 

The package supports time series aggregation for energy systems time series datasets. 

The package complies with the guideline, has been published in the journal of open source software, has been contributed to by authors beyond the core authors, is structured so that it can be reused and extended (including that we provide comments and documentation).